+++ lib/javafx-sdk-17.0.1/lib/javafx.properties
@@ -1,0 +1,3 @@
+javafx.version=17.0.1
+javafx.runtime.version=17.0.1+1
+javafx.runtime.build=1
