+++ src/main/Main.java
@@ -1,0 +1,24 @@
+package src.main;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class Main  extends Application {
+    @Override
+    public void start(Stage stage) throws IOException {
+        FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("../views/mainScreen.fxml"));
+        Scene scene = new Scene(fxmlLoader.load(), 900, 600);
+        stage.setResizable(false);
+        stage.setTitle("Beemail");
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    public static void main(String[] args) {
+        launch();
+    }
+}
